Move workflow/destination lag metrics out of global metrics file
