+++ plugin-server/src/cdp/services/native-destination-executor.service.ts
@@ -1,6 +1,5 @@
 import { Histogram } from 'prom-client'
 
-import { destinationE2eLagMsSummary } from '~/main/ingestion-queues/metrics'
 import { PluginsServerConfig } from '~/types'
 import { parseJSON } from '~/utils/json-parse'
 
@@ -8,6 +7,7 @@ import { logger } from '../../utils/logger'
 import { FetchOptions, FetchResponse } from '../../utils/request'
 import { NATIVE_HOG_FUNCTIONS_BY_ID } from '../templates'
 import { CyclotronJobInvocationHogFunction, CyclotronJobInvocationResult, Response } from '../types'
+import { destinationE2eLagMsSummary } from '../utils'
 import { CDP_TEST_ID, createAddLogFunction, isNativeHogFunction } from '../utils'
 import { createInvocationResult } from '../utils/invocation-utils'
 import { cdpTrackedFetch, getNextRetryTime, isFetchResponseRetriable } from './hog-executor.service'
